lint + cleanup after merge
