+++ src/App/AppWrapper.tsx
@@ -5,10 +5,9 @@ import { useDetectMode, useDisableRefresh, useLangaugeResolver, useListenForExte
 
 import { fetchAppConfig } from '../store/useAppConfigStore.ts';
 import setupLogger from '../utils/shared-logger.ts';
-import App from './App.tsx';
 import useListenForCriticalProblem from '@app/hooks/useListenForCriticalProblem.tsx';
-
-import { setMiningNetwork } from '@app/store/useMiningStore.ts';
+import { setMiningNetwork } from '@app/store/miningStoreActions.ts';
+import App from './App.tsx';
 
 // FOR ANYTHING THAT NEEDS TO BE INITIALISED
